+++ next.config.js
@@ -1,4 +1,11 @@
 /** @type {import('next').NextConfig} */
 module.exports = {
     reactStrictMode: true,
+    redirects: () => [
+        {
+            source: "/mailinglist",
+            destination: "/newsletter",
+            permanent: true,
+        },
+    ],
 }