+++ pages/kickstarter/index.tsx
@@ -1,6 +1,4 @@
 import { NextPage } from "next"
 import RedirectPage from "~/components/RedirectPage"
-const Redirect: NextPage = () => (
-    <RedirectPage href="//www.kickstarter.com/projects/keesey/paleocene-5-comic-book?ref=56eiuo" title="Kickstarter" />
-)
+const Redirect: NextPage = () => <RedirectPage href="//www.kickstarter.com/projects/keesey" title="Kickstarter" />
 export default Redirect